Call a renewal callback on token expiry
In Partner apps tokens can be auto-renewed and it's easier to put
the renewal code in one place and have Xeroizer auto-retry than
catch Xeroizer::OAuth::TokenExpired all over the place.

This commit adds a hook which calls user code with an identifier so
that users of Xeroizer can attempt to renew the token and have
Xeroizer immediately try the call again.

The identifier can be anything, it's passed through to the lambda.

Example:

identifier = organisation.id
@client = Xeroizer::PartnerApplication.new(
  ..
  after_token_expired: -> (identifier, client) {
    org = Org.find(identifier)
    client.renew_access_token(org.xero_access_token, org.xero_access_secret, org.xero_session_handle)

    org.xero_access_token = client.access_token.token
    org.xero_access_key = client.access_token.secret
    org.xero_session_handle = client.session_handle

    org.save!
  }
)